Update ProjectRequestForm.js
Discovered the trailing comma is required when filtering the DeptIDs list column. Otherwise filtering on DeptIDs CONTAINS 1 will return 1, 13, 21, and anything else that contains the number 1. By using the comma in the filter "1," we're able to specify individual IDs; however, the last number in the string will never be returned if the trailing comma is removed.