Create TiedProjects.js
JavaScript file included in Nintex form for BrightWork Tied Projects list.
Script queries the current site's Project Statement for Project Number, uses that to query the site collection's Project Cache for Parent Site ID, and uses that to query all sibling sites, excluding the current site.
That list of sites is then filled into a drop down choice field on the form, which stores the chosen site to the list.