Complete Overhaul
Added features to limit the returned sites by sites already listed.
Also added function to set form variables when a site is selected from the drop down so the correct data is passed along to the list columns.
Added global variables so the values are available throughout the entire script.
Included variable to track which function the script is currently in, and output that value to the console when the query failed function runs.
Increased the data output to the console for added troubleshooting assistance.